fixing null errors

+++ ft_strlen.c
@@ -1,6 +1,8 @@
-int ft_strlen(char *c)
+#include "libft.h"
+
+size_t ft_strlen(char *c)
 {
-    int i;
+    size_t i;
     i = 0;
     while (c[i] != '\0')
         i++;
